# docs/swagger.yaml
openapi: 3.0.0
info:
  title: LMS API
  description: API documentation for learning management system with courses, videos, authentication, and progress tracking
  version: 1.0.0
servers:
  - url: /api
    description: Development server
...
paths:
...
  /admin/progress-review:
    get:
      tags:
        - Admin
      summary: Get all user progress
      description: Admin endpoint to review progress of all users across all courses
      security:
        - BearerAuth: []
        - SessionAuth: []
      responses:
        '200':
          description: List of all user progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProgress'
        '500':
          description: Failed to fetch user progress
  /admin/grant-access:
    post:
      tags:
        - Admin
      summary: Grant course access
      description: Admin endpoint to grant a user access to a specific course
      security:
        - BearerAuth: []
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - courseName
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                courseName:
                  type: string
                  example: "Introduction to JavaScript"
      responses:
        '200':
          description: Access granted successfully
        '404':
          description: User or course not found
        '500':
          description: Error granting access
  
  # Course Routes
  /course:
    get:
      tags:
        - Courses
      summary: Get all courses
      description: Retrieves a list of all available courses
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '500':
          description: Server error
...
  /video:
    get:
      tags:
        - Videos
      summary: Get all videos
      description: Retrieves a list of all videos
      responses:
        '200':
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Video'
        '500':
          description: Server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
  schemas:
...
    Course:
      type: object
      properties:
        _id:
          type: string
          example: "60d21b4667d0d8992e610c86"
        title:
          type: string
          example: "JavaScript Fundamentals"
        description:
          type: string
          example: "Learn the basics of JavaScript programming"
        videos:
          type: array
          items:
            $ref: '#/components/schemas/VideoObject'
        totalMinutes:
          type: number
          example: 120
    VideoObject:
      type: object
      properties:
        title:
          type: string
          example: "Introduction to JavaScript"
        description:
          type: string
          example: "An overview of JavaScript basics"
        url:
          type: string
          example: "/uploads/1620123456789-123456789.mp4"
        duration:
          type: number
          example: 20
    Video:
      type: object
      properties:
        _id:
          type: string
          example: "60d21b4667d0d8992e610c87"
        title:
          type: string
          example: "Introduction to JavaScript"
        description:
          type: string
          example: "An overview of JavaScript basics"
        url:
          type: string
          example: "/uploads/1620123456789-123456789.mp4"
...
    UserProgress:
      type: object
      properties:
        _id:
          type: string
          example: "60d21b4667d0d8992e610c89"
        userId:
          type: string
          example: "60d21b4667d0d8992e610c85"
        courseId:
          type: string
          example: "60d21b4667d0d8992e610c86"
        totalMinutes:
          type: number
          example: 120
        completedMinutes:
          type: number
          example: 45
        progress:
          type: number
          example: 38